+++ src/ringbuffer.h
@@ -38,29 +38,29 @@ namespace ClickTrack{
              * range of the buffer. If you try to access a time point not
              * available in the current range, throws exception.
              */
-            SampleT get(const unsigned t);
+            SampleT get(const unsigned long t);
 
             /* Copies a range of values into a provided buffer. If you try to
              * access any time points not available in the current range, throws
              * exception.
              */
-            void get_range(std::vector<SampleT>& buffer, const unsigned start,
-                    const unsigned end);
+            void get_range(std::vector<SampleT>& buffer, 
+                    const unsigned long start, const unsigned long end);
 
             /* Adds a sample as the next time step in the buffer. May overwrite
              * the oldest time step. Returns the timestamp of the added value.
              */
-            unsigned add(SampleT s);
+            unsigned long add(SampleT s);
 
             /* Exposes a reference to an element in the buffer so that you can
              * manually write to and modify a buffer point.
              */
-            SampleT& operator[] (unsigned t);
+            SampleT& operator[] (unsigned long t);
 
             /* Getters to expose the lowest and high timestamp
              */
-            unsigned get_lowest_timestamp();
-            unsigned get_highest_timestamp();
+            unsigned long get_lowest_timestamp();
+            unsigned long get_highest_timestamp();
 
             /* WARNING: This function should not be used in ordinary
              * circumstances. It exists only for adding a new buffer during
@@ -69,11 +69,11 @@ namespace ClickTrack{
             void set_new_startpoint(unsigned t);
 
         private:
-            unsigned start_t; // earliest time still in the buffer
-            unsigned end_t;   // latest time still in the buffer
-            unsigned size;    // number of samples currently in buffer
+            unsigned long start_t; // earliest time still in the buffer
+            unsigned long end_t;   // latest time still in the buffer
+            unsigned long size;    // number of samples currently in buffer
 
-            unsigned buffer_size;  // the number of elements in the buffer
+            unsigned long buffer_size;  // the number of elements in the buffer
             std::vector<SampleT> samples; // the actual ring array
     };
 }
